HEEDLS-621 Update links on searchable elements with a return page number in js
